refactor: notice 패키지 구조 개편 및 AiStrategy FK 변경
- notice 하위 sub-package → entity/, repository/ 통합
- 제거: NoticeCompanyInfo, NoticeGuideline (dead entity)
- 제거: ApplicationQuestion, QuestionType (CoverLetterItem으로 대체)
- AiStrategy FK: ApplicationQuestion → CoverLetterItem
- NoticeService import 경로 업데이트
- CoverLetterItem, CoverLetterItemService import 경로 업데이트

closes #61 (partial)
