+++ src/main/java/back/pickd/notice/document/ApplicationDocumentRepository.java
@@ -1,8 +1,1 @@
-package back.pickd.notice.document;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ApplicationDocumentRepository extends JpaRepository<ApplicationDocument, Long> {
-}
+// MOVED: notice/entity/ 또는 notice/repository/ 로 이동됨
