+++ src/main/java/back/pickd/notice/enums/QuestionType.java
@@ -1,8 +1,2 @@
-package back.pickd.notice.enums;
-
-public enum QuestionType {
-    COVER_LETTER,    // 자기소개서
-    FREE_FORM,       // 자유 형식
-    JOB_DESCRIPTION, // 직무 기술서
-    ADDITIONAL       // 추가 문항
-}
+// REMOVED: ApplicationQuestion 제거로 함께 삭제
+// 삭제 예정: git rm src/main/java/back/pickd/notice/enums/QuestionType.java
