+++ src/main/java/back/pickd/notice/notice/NoticeRepository.java
@@ -1,16 +1,1 @@
-package back.pickd.notice.notice;
-
-import back.pickd.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface NoticeRepository extends JpaRepository<Notice, Long> {
-
-    List<Notice> findAllByUserOrderByIdDesc(User user);
-
-    Optional<Notice> findByIdAndUser(Long id, User user);
-}
+// MOVED: notice/entity/ 또는 notice/repository/ 로 이동됨
