+++ src/main/java/back/pickd/notice/preference/SectionPreferenceRepository.java
@@ -1,8 +1,1 @@
-package back.pickd.notice.preference;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface SectionPreferenceRepository extends JpaRepository<SectionPreference, Long> {
-}
+// MOVED: notice/entity/ 또는 notice/repository/ 로 이동됨
