+++ src/main/java/back/pickd/notice/qualification/SectionQualificationRepository.java
@@ -1,8 +1,1 @@
-package back.pickd.notice.qualification;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface SectionQualificationRepository extends JpaRepository<SectionQualification, Long> {
-}
+// MOVED: notice/entity/ 또는 notice/repository/ 로 이동됨
