+++ src/main/java/back/pickd/notice/repository/ApplicationDocumentRepository.java
@@ -1,0 +1,9 @@
+package back.pickd.notice.repository;
+
+import back.pickd.notice.entity.ApplicationDocument;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ApplicationDocumentRepository extends JpaRepository<ApplicationDocument, Long> {
+}
