+++ src/main/java/back/pickd/notice/repository/NoticeProcessRepository.java
@@ -1,0 +1,9 @@
+package back.pickd.notice.repository;
+
+import back.pickd.notice.entity.NoticeProcess;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface NoticeProcessRepository extends JpaRepository<NoticeProcess, Long> {
+}
