+++ src/main/java/back/pickd/notice/repository/NoticeSectionRepository.java
@@ -1,0 +1,9 @@
+package back.pickd.notice.repository;
+
+import back.pickd.notice.entity.NoticeSection;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface NoticeSectionRepository extends JpaRepository<NoticeSection, Long> {
+}
