+++ src/main/java/back/pickd/notice/repository/SectionPreferenceRepository.java
@@ -1,0 +1,9 @@
+package back.pickd.notice.repository;
+
+import back.pickd.notice.entity.SectionPreference;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SectionPreferenceRepository extends JpaRepository<SectionPreference, Long> {
+}
