+++ src/main/java/back/pickd/notice/section/NoticeSectionRepository.java
@@ -1,8 +1,1 @@
-package back.pickd.notice.section;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface NoticeSectionRepository extends JpaRepository<NoticeSection, Long> {
-}
+// MOVED: notice/entity/ 또는 notice/repository/ 로 이동됨
